Move SelectorMediaSourceTester to domain layer
