+++ app/shared/ui-settings/src/commonMain/kotlin/ui/settings/mediasource/RefreshIndicatedHeadlineRow.kt
@@ -39,48 +39,13 @@ import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import me.him188.ani.app.data.repository.RepositoryAuthorizationException
-import me.him188.ani.app.data.repository.RepositoryException
 import me.him188.ani.app.data.repository.RepositoryNetworkException
 import me.him188.ani.app.data.repository.RepositoryRateLimitedException
 import me.him188.ani.app.data.repository.RepositoryServiceUnavailableException
 import me.him188.ani.app.data.repository.RepositoryUnknownException
+import me.him188.ani.app.domain.mediasource.test.RefreshResult
 import me.him188.ani.app.ui.foundation.widgets.LocalToaster
 
-/**
- * 标记接口, 表示一个通用的刷新结果.
- */
-interface RefreshResult {
-    /**
-     * 刷新成功, 不会显示错误按钮.
-     */
-    interface Success : RefreshResult
-
-    interface InProgress : RefreshResult
-
-    /**
-     * 刷新失败, 会显示错误按钮.
-     */
-    sealed interface Failed : RefreshResult
-
-    /**
-     * 一个已知类型的 API 错误
-     */
-    interface ApiError : Failed {
-        val exception: RepositoryException
-    }
-
-    /**
-     * 配置有误, 例如测试 RSS 数据源. UI 只会显示 "配置不完整". 你需要让你的编辑框自己显示 error.
-     */
-    interface InvalidConfig : Failed
-
-    /**
-     * 一个任意类型异常. 这属于不期望遇到的错误 (bug).
-     */
-    interface UnknownError : Failed {
-        val exception: Throwable
-    }
-}
 
 /**
  * 包含一个 Text, 一个刷新按钮, 一个错误提示的 [Row].
